refactoring Modal code
