+++ src/app/[locale]/layout.jsx
@@ -1,5 +1,5 @@
 import CookiesModal from "@/src/components/CookiesModal/CookiesModal";
-import PaymentModal from "@/src/components/PaymentModal/PaymentModal";
+import PaymentModal from "@/src/components/modals/PaymentModal/PaymentModal";
 import Footer from "@/src/components/shared/Footer/Footer";
 import Header from "@/src/components/shared/Header/Header";
 import { NextIntlClientProvider } from "next-intl";
