+++ src/app/[locale]/ui-kit/page.jsx
@@ -3,7 +3,7 @@ import FooterLinks from "@/src/components/shared/FooterLinks/FooterLinks";
 import PartnerCard from "@/src/components/shared/PartnerCard/PartnerCard";
 import SocialIcons from "@/src/components/shared/SocialIcons/SocialIcons";
 import HeroCard from "@/src/components/shared/HeroCard/HeroCard";
-import RegistrationFormModal from "@/src/components/RegistrationFormModal/RegistrationFormModal";
+import RegistrationFormModal from "@/src/components/modals/RegistrationFormModal/RegistrationFormModal";
 import ControlBtnRegistrationFormModal from "@/src/components/shared/controlButtons/ControlBtnRegistrationFormModal";
 import JoinProjectSection from "@/src/components/JoinProjectSection/JoinProjectSection";
