+++ src/components/modals/LayoutModal/LayoutModal.jsx
@@ -2,9 +2,9 @@
 import { useEffect, useCallback } from "react";
 import { createPortal } from "react-dom";
 import { useBodyLock } from "@/src/lib/hooks/useBodyLock";
-import styles from "./Modal.module.scss";
+import styles from "./LayoutModal.module.scss";
 
-const Modal = ({ children, isOpen, handleClose }) => {
+const LayoutModal = ({ children, isOpen, handleClose }) => {
   const closeOnEscape = useCallback(
     (e) => {
       if (e.key === "Escape") {
@@ -31,4 +31,4 @@ const Modal = ({ children, isOpen, handleClose }) => {
   );
 };
 
-export default Modal;
+export default LayoutModal;
