+++ src/components/modals/PaymentModal/FormPayment/FormPayment.module.scss
@@ -1,5 +1,5 @@
-@use "../../../styles/variables" as var;
-@use "../../../styles/mixins.scss" as mx;
+@use "../../../../styles/variables" as var;
+@use "../../../../styles/mixins.scss" as mx;
 
 .form{
   color: var.$secondary-white;
