+++ src/components/modals/PaymentModal/ThanksCard/ThanksCard.module.scss
@@ -1,5 +1,5 @@
-@use "../../../styles/variables" as var;
-@use "../../../styles/mixins.scss" as mx;
+@use "../../../../styles/variables" as var;
+@use "../../../../styles/mixins.scss" as mx;
 
 .thanks{
   display: flex;
