+++ src/components/modals/RegistrationFormModal/RegistrationFormModal.jsx
@@ -1,17 +1,16 @@
 "use client";
 import React from "react";
-//import ReactDOM from "react-dom";
 import styles from "./RegistrationFormModal.module.scss";
-import InputField from "../shared/InputField/InputField";
 import { useForm } from "react-hook-form";
 import { useTranslations } from "next-intl";
 import { formScheme } from "./formScheme";
-import MainButton from "../shared/MainButton/MainButton";
-
-import Modal from "../shared/Modal/Modal";
 import stateRegistrationFormModal from "@/src/state/stateRegistrationFormModal";
 import { useBodyLock } from "@/src/lib/hooks/useBodyLock";
-import CloseBtn from "../shared/CloseBtn/CloseBtn";
+import Modal from "../LayoutModal/LayoutModal";
+import CloseBtn from "../../shared/CloseBtn/CloseBtn";
+import MainButton from "../../shared/MainButton/MainButton";
+import InputField from "../../shared/InputField/InputField";
+import LayoutModal from "../LayoutModal/LayoutModal";
 
 const RegistrationFormModal=()=>{
   const t = useTranslations("Modal_form");
@@ -38,10 +37,10 @@ const RegistrationFormModal=()=>{
   };
   const optionsSpec = [
     { id: "uiuxDesigner", label: "UI/UX Designer" },
-    { id: "backend", label: "Backend" },
-    { id: "qaManualEngineer", label: "QA Manual Engineer" },
     { id: "frontend", label: "Frontend" },
+    { id: "backend", label: "Backend" },
     { id: "fullstackEngineer", label: "Fullstack Engineer" },
+    { id: "qaManualEngineer", label: "QA Manual Engineer" },
     { id: "projectManager", label: "Project Manager" },
   ];
 
@@ -85,7 +84,9 @@ const RegistrationFormModal=()=>{
     {
       id: "email",
       placeholder: "email@gmail.com",
-      registerOptions: register("email", { ...formScheme.email }),
+      registerOptions: register("email", { ...formScheme.email, onBlur:() => {
+        trigger("email")
+      } }),
       isError: !!errors.email,
       version: "input",
       label: t("email"),
@@ -93,7 +94,9 @@ const RegistrationFormModal=()=>{
     {
       id: "phone",
       placeholder: "+380 xxx xxx xxx",
-      registerOptions: register("phone", { ...formScheme.phone }),
+      registerOptions: register("phone", { ...formScheme.phone, onBlur:() => {
+        trigger("phone")
+      } }),
       isError: !!errors.phone,
       version: "input",
       label: t("phone"),
@@ -101,7 +104,9 @@ const RegistrationFormModal=()=>{
     {
       id: "discord",
       placeholder: t("discord"),
-      registerOptions: register("discord", { ...formScheme.discord }),
+      registerOptions: register("discord", { ...formScheme.discord, onBlur:() => {
+        trigger("discord")
+      } }),
       isError: !!errors.discord,
       version: "input",
       label: t("discord"),
@@ -109,7 +114,9 @@ const RegistrationFormModal=()=>{
     {
       id: "linkedin",
       placeholder: t("linkedin_placeholder"),
-      registerOptions: register("linkedin", { ...formScheme.linkedin }),
+      registerOptions: register("linkedin", { ...formScheme.linkedin, onBlur:() => {
+        trigger("linkedin")
+      } }),
       isError: !!errors.linkedin,
       version: "input",
       label: t("linkedin"),
@@ -131,7 +138,7 @@ const RegistrationFormModal=()=>{
   ];
 
   return (
-    <Modal isOpen handleClose={onClose}>
+    <LayoutModal isOpen handleClose={onClose}>
       <div className={styles.wrapper} onClick={(e) => {
         onClose()
         e.stopPropagation()
@@ -148,13 +155,13 @@ const RegistrationFormModal=()=>{
               {inputFields.map((field) => (
                 <div key={field.id}>
                   <InputField {...field} />
-                  {field.isError&&<p>{t(`error_message.${field.id}`)}</p>}
+                  {field.isError && <p>{t(`error_message.${field.id}`)}</p>}
                 </div>
               ))}
             </ul>
             <MainButton
               type="submit"
-              disabled={!isDirty || !isValid}
+             // disabled={!isDirty || !isValid}
               className={styles.submit}
               //variant={"modal"}
             >
@@ -163,7 +170,7 @@ const RegistrationFormModal=()=>{
           </form>
         </div>
       </div>
-    </Modal>
+    </LayoutModal>
   );
 
 };
