+++ src/components/modals/RegistrationFormModal/RegistrationFormModal.module.scss
@@ -1,5 +1,5 @@
-@use "../../styles/mixins" as mix;
-@use "../../styles/variables" as var;
+@use "../../../styles/mixins" as mix;
+@use "../../../styles/variables" as var;
 
 .wrapper{
   width: 100%;
